Use @actions/glob for pattern matching
