+++ src/save.ts
@@ -4,7 +4,6 @@ import * as cacheHttpClient from "./cacheHttpClient";
 import { Events, Inputs, State } from "./constants";
 import { createTar } from "./tar";
 import * as utils from "./utils/actionUtils";
-import * as pathUtils from "./utils/pathUtils";
 
 async function run(): Promise<void> {
     try {
@@ -44,7 +43,7 @@ async function run(): Promise<void> {
             return;
         }
         core.debug(`Cache ID: ${cacheId}`);
-        const cachePaths = pathUtils.expandPaths(
+        const cachePaths = await utils.expandPaths(
             core
                 .getInput(Inputs.Path)
                 .split("\n")
